+++ lib/cargoBench.nix
@@ -1,0 +1,18 @@
+{ mkCargoDerivation
+}:
+
+{ cargoExtraArgs ? ""
+, cargoBenchExtraArgs ? ""
+, ...
+}@origArgs:
+let
+  args = builtins.removeAttrs origArgs [
+    "cargoExtraArgs"
+    "cargoBenchExtraArgs"
+  ];
+in
+mkCargoDerivation (args // {
+  pnameSuffix = "-bench";
+
+  buildPhaseCargoCommand = "cargoWithProfile bench ${cargoExtraArgs} ${cargoBenchExtraArgs}";
+})