+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* `cargoBench` can now be used to run cargo benchmarks
+
 ### Changed
 * **Breaking**: dropped compatibility for Nix versions below 2.18.2
 * **Breaking**: dropped compatibility for nixpkgs-23.11.